Fix MOTM page syntax
